Changed Key Strategy for operator. Make sure the model is consistent.

+++ src/main/java/operators/events/WatermarkEvent.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 public class WatermarkEvent extends IterableOperatorEvent {
     public long timestamp;
 
-    public WatermarkEvent(long timestamp, Short currentIteration) {
+    public WatermarkEvent(long timestamp, Byte currentIteration) {
         super(currentIteration);
         this.timestamp = timestamp;
     }
